+++ pkg/apis/deployment/v1/deployment_status.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2016-2022 ArangoDB GmbH, Cologne, Germany
+// Copyright 2016-2023 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -196,4 +196,20 @@ func (ds *DeploymentStatus) UpdateServerGroupStatus(group ServerGroup, gspec Ser
 	case ServerGroupSyncWorkers:
 		ds.SyncWorkers = &gspec
 	}
+}
+
+// IsHibernated returns two boolean values:
+// - first bool describes if hibernation is turned on.
+// - second bool describes if hibernation is done.
+func (ds *DeploymentStatus) IsHibernated() (bool, bool) {
+	c, ok := ds.Conditions.Get(ConditionTypeHibernation)
+	if !ok {
+		return false, false
+	}
+
+	if c.IsTrue() {
+		return true, true
+	}
+
+	return true, false
 }