+++ pkg/apis/deployment/v2alpha1/conditions.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2023 ArangoDB GmbH, Cologne, Germany
+// Copyright 2016-2023 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -126,6 +126,9 @@ const (
 
 	// ConditionTypeSyncEnabled Define if sync is enabled
 	ConditionTypeSyncEnabled ConditionType = "SyncEnabled"
+
+	// ConditionTypeHibernation should be set when deployment is in hibernation mode.
+	ConditionTypeHibernation ConditionType = "Hibernation"
 )
 
 // Condition represents one current condition of a deployment or deployment member.