+++ pkg/deployment/context_impl.go
@@ -387,23 +387,6 @@ func (d *Deployment) DeletePod(ctx context.Context, podName string, options meta
 	return nil
 }
 
-// CleanupPod deletes a given pod with force and explicit UID.
-// If the pod does not exist, the error is ignored.
-func (d *Deployment) CleanupPod(ctx context.Context, p *core.Pod) error {
-	log := d.log
-	podName := p.GetName()
-	options := meta.NewDeleteOptions(0)
-	options.Preconditions = meta.NewUIDPreconditions(string(p.GetUID()))
-	err := globals.GetGlobalTimeouts().Kubernetes().RunWithTimeout(ctx, func(ctxChild context.Context) error {
-		return d.PodsModInterface().Delete(ctxChild, podName, *options)
-	})
-	if err != nil && !kerrors.IsNotFound(err) {
-		log.Err(err).Str("pod", podName).Debug("Failed to cleanup pod")
-		return errors.WithStack(err)
-	}
-	return nil
-}
-
 // RemovePodFinalizers removes all the finalizers from the Pod with given name in the namespace
 // of the deployment. If the pod does not exist, the error is ignored.
 func (d *Deployment) RemovePodFinalizers(ctx context.Context, podName string) error {