+++ pkg/deployment/member/state.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2016-2022 ArangoDB GmbH, Cologne, Germany
+// Copyright 2016-2023 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -120,6 +120,10 @@ func (s *stateInspector) RefreshState(ctx context.Context, members api.Deploymen
 	servingGroup := mode.ServingGroup()
 	var client driver.Client
 	members.ForEach(func(id int) {
+		if members[id].Member.Phase == api.MemberPhaseHibernated {
+			return
+		}
+
 		ctxChild, cancel := globals.GetGlobalTimeouts().ArangoDCheck().WithTimeout(ctx)
 		defer cancel()