+++ pkg/deployment/reconcile/action_hibernate.go
@@ -21,27 +21,15 @@
 package reconcile
 
 import (
-	"context"
-
 	api "github.com/arangodb/kube-arangodb/pkg/apis/deployment/v1"
 )
 
-const (
-	// Component name for reconciliation of this package
-	reconciliationComponent = "deployment_reconciliation"
-)
+// newHibernateAction creates a new Action that implements the given
+// planned hibernation action.
+func newHibernateAction(action api.Action, actionCtx ActionContext) Action {
+	a := &actionHibernate{}
 
-const (
-	BackOffCheck  api.BackOffKey = "check"
-	LicenseCheck  api.BackOffKey = "license"
-	TimezoneCheck api.BackOffKey = "timezone"
-)
+	a.actionImpl = newActionImplDefRef(action, actionCtx)
 
-// CreatePlan considers the current specification & status of the deployment creates a plan to
-// get the status in line with the specification.
-// If a plan already exists, nothing is done.
-func (d *Reconciler) CreatePlan(ctx context.Context) (error, bool) {
-	return d.generatePlan(ctx, d.generatePlanFunc(d.createHighPlan, plannerHigh{}),
-		d.generatePlanFunc(d.createResourcesPlan, plannerResources{}),
-		d.generatePlanFunc(d.createNormalPlan, plannerNormal{}))
+	return a
 }